+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
--
+- Support passing `PrecisionPlugin` instances to the `precision` flag ([#12540](https://github.com/PyTorchLightning/pytorch-lightning/pull/12540))
 
 
 -